misc updates

+++ include/ray/intersect.cuh
@@ -36,7 +36,7 @@ template <bool normal_and_uv>
 __inline__ __host__
     __device__ thrust::optional<BestIntersectionGeneral<normal_and_uv>>
     get_shape_intersection(Span<const scene::ShapeData> shapes,
-                           const uint16_t shape_idx,
+                           const unsigned shape_idx,
                            const Eigen::Vector3f &world_space_eye,
                            const Eigen::Vector3f &world_space_direction) {
   const auto &shape = shapes[shape_idx];
@@ -97,10 +97,10 @@ __inline__ __host__ __device__ void solve_general_intersection(
     auto inv_direction = (1.0f / world_space_direction.array()).eval();
 
     struct StackData {
-      uint16_t node_index;
+      unsigned node_index;
       uint8_t depth;
 
-      __device__ __host__ StackData(uint16_t node_index, uint8_t depth)
+      __device__ __host__ StackData(unsigned node_index, uint8_t depth)
           : node_index(node_index), depth(depth) {}
 
       __device__ __host__ StackData() {}
@@ -111,8 +111,8 @@ __inline__ __host__ __device__ void solve_general_intersection(
       node_stack[0] = StackData(kdtree_nodes.nodes.size() - 1, 0);
       uint8_t node_stack_size = 1;
 
-      thrust::optional<std::array<uint16_t, 2>> start_end = thrust::nullopt;
-      uint16_t current_shape_index = 0;
+      thrust::optional<std::array<unsigned, 2>> start_end = thrust::nullopt;
+      unsigned current_shape_index = 0;
 
       while (node_stack_size != 0 || start_end.has_value()) {
         while (!start_end.has_value() && node_stack_size != 0) {
@@ -149,7 +149,7 @@ __inline__ __host__ __device__ void solve_general_intersection(
                   node_stack[node_stack_size - 1] = StackData(first, new_depth);
                   node_stack_size++; // counter act --;
                 },
-                [&](const std::array<uint16_t, 2> &data) {
+                [&](const std::array<unsigned, 2> &data) {
                   start_end = thrust::make_optional(data);
                   current_shape_index = data[0];
                 });
@@ -160,7 +160,7 @@ __inline__ __host__ __device__ void solve_general_intersection(
 
         if (start_end.has_value()) {
           auto local_end_shape = (*start_end)[1];
-          uint16_t shape_idx = current_shape_index;
+          unsigned shape_idx = current_shape_index;
           current_shape_index++;
           if (current_shape_index >= local_end_shape) {
             start_end = thrust::nullopt;
@@ -173,7 +173,7 @@ __inline__ __host__ __device__ void solve_general_intersection(
       }
     }
   } else {
-    for (uint16_t shape_idx = 0; shape_idx < 0 + kdtree_nodes.num_shape;
+    for (unsigned shape_idx = 0; shape_idx < 0 + kdtree_nodes.num_shape;
          shape_idx++) {
       if (solve_index(shape_idx)) {
         return;
